pyproject.toml
[project]
name = "labresta-sync"
version = "0.1.0"
description = "Automated price and availability sync from supplier XML feeds to prom.ua"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1",
    "flask-sqlalchemy>=3.1",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "lxml>=5.0",
    "requests>=2.32",
    "python-dotenv>=1.0",
    "openpyxl>=3.1",
    "chardet>=5.0",
    "tenacity>=8.0",
    "rapidfuzz>=3.0",
    "flask-apscheduler>=1.13.1",
    "flask-login>=0.6",
    "flask-wtf>=1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-flask>=1.3",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"
